delete user functionallity added

+++ src/users/users.service.ts
@@ -1,5 +1,7 @@
-import { Injectable } from "@nestjs/common";
+import { Injectable, UnauthorizedException } from "@nestjs/common";
+import { JwtService } from "@nestjs/jwt";
 import { InjectRepository } from "@nestjs/typeorm";
+import { JwtPayload } from "src/data/dto/jwt-payload.dto";
 import { User } from "src/data/entity/user.entity";
 import { Repository } from "typeorm";
 
@@ -8,6 +10,7 @@ export class UsersService {
   constructor(
     @InjectRepository(User)
     private userRepository: Repository<User>,
+    private jwtService: JwtService,
   ) {}
 
   async findAll(): Promise<User[]> {
@@ -38,8 +41,16 @@ export class UsersService {
     });
   }
 
-  async delete(id: string): Promise<void> {
-    console.log(id);
+  async delete(id: string, authorisation: string): Promise<void> {
+    const { email } = this.jwtService.decode<JwtPayload>(authorisation);
+    const crtUser = await this.userRepository.findOne({
+      where: {
+        email: email,
+      },
+    });
+    if (crtUser?.id.toString() != id) {
+      throw new UnauthorizedException("User is unauthorised");
+    }
     this.userRepository.delete(id);
   }
 }
